requirements.txt
pytest
allure-pytest==2.9.45
selene==2.0.0b5